deploy_app/webapp/requirements.txt
python-multipart
openpyxl
fastapi
uvicorn[standard]
pydantic
pyngrok
joblib
pandas
streamlit
transformers
Pillow
python-multipart